extracted classes and enums to it's own files. Moved shutdown hook to cyphershell
