+++ cypher-shell/src/main/java/org/neo4j/shell/cli/FailBehavior.java
@@ -1,0 +1,6 @@
+package org.neo4j.shell.cli;
+
+public enum FailBehavior {
+    FAIL_FAST,
+    FAIL_AT_END
+}
