+++ cypher-shell/src/main/java/org/neo4j/shell/prettyprint/StatisticsCollector.java
@@ -3,7 +3,7 @@ package org.neo4j.shell.prettyprint;
 import org.neo4j.driver.v1.StatementResult;
 import org.neo4j.driver.v1.summary.ResultSummary;
 import org.neo4j.driver.v1.summary.SummaryCounters;
-import org.neo4j.shell.cli.CliArgHelper;
+import org.neo4j.shell.cli.Format;
 
 import javax.annotation.Nonnull;
 import java.util.ArrayList;
@@ -11,14 +11,14 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class StatisticsCollector {
-    private CliArgHelper.Format format;
+    private Format format;
 
-    public StatisticsCollector(@Nonnull CliArgHelper.Format format) {
+    public StatisticsCollector(@Nonnull Format format) {
         this.format = format;
     }
 
     public String collect(@Nonnull StatementResult result) {
-        if (CliArgHelper.Format.VERBOSE == format) {
+        if (Format.VERBOSE == format) {
             return collectStatistics(result);
         } else {
             return "";
