+++ cypher-shell/src/test/java/org/neo4j/shell/cli/CliArgHelperTest.java
@@ -13,19 +13,19 @@ public class CliArgHelperTest {
 
     @Test
     public void testFailFastIsDefault() {
-        assertEquals("Unexpected fail-behavior", CliArgHelper.FailBehavior.FAIL_FAST,
+        assertEquals("Unexpected fail-behavior", FailBehavior.FAIL_FAST,
                 CliArgHelper.parse(asArray()).getFailBehavior());
     }
 
     @Test
     public void testFailFastIsParsed() {
-        assertEquals("Unexpected fail-behavior", CliArgHelper.FailBehavior.FAIL_FAST,
+        assertEquals("Unexpected fail-behavior", FailBehavior.FAIL_FAST,
                 CliArgHelper.parse(asArray("--fail-fast")).getFailBehavior());
     }
 
     @Test
     public void testFailAtEndIsParsed() {
-        assertEquals("Unexpected fail-behavior", CliArgHelper.FailBehavior.FAIL_AT_END,
+        assertEquals("Unexpected fail-behavior", FailBehavior.FAIL_AT_END,
                 CliArgHelper.parse(asArray("--fail-at-end")).getFailBehavior());
     }
